refactor chart tests
